Moved userMeta to external file
